Allow pressing escape to skip splash screen
